# docs/openapi.json
{
  "components": {
    "schemas": {
      "CompareMetricRow": {
        "properties": {
          "display_values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Display Values",
            "type": "object"
          },
          "metric_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric Category"
          },
          "metric_name": {
            "title": "Metric Name",
            "type": "string"
          },
          "metric_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric Unit"
          },
          "values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "metric_name",
          "metric_unit",
          "metric_category",
          "values",
          "display_values"
        ],
        "title": "CompareMetricRow",
        "type": "object"
      },
...
      "CompareResponse": {
        "properties": {
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/CompareMetricRow"
            },
            "title": "Metrics",
            "type": "array"
          },
          "model_outputs": {
            "items": {
              "$ref": "#/components/schemas/CompareModelRow"
            },
            "title": "Model Outputs",
            "type": "array"
          },
          "overviews": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SummaryResponse"
            },
            "title": "Overviews",
            "type": "object"
          },
          "region_ids": {
            "items": {
              "type": "string"
            },
            "title": "Region Ids",
            "type": "array"
          },
          "regions": {
            "items": {
              "$ref": "#/components/schemas/Region"
            },
            "title": "Regions",
            "type": "array"
          }
        },
        "required": [
          "region_ids",
          "regions",
          "metrics",
          "model_outputs",
          "overviews"
        ],
        "title": "CompareResponse",
        "type": "object"
      },
      "Evidence": {
        "properties": {
          "evidence_id": {
            "title": "Evidence Id",
            "type": "integer"
          },
          "evidence_text": {
            "title": "Evidence Text",
            "type": "string"
          },
          "evidence_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Type"
          },
          "region_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Id"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          }
        },
        "required": [
          "evidence_id",
          "region_id",
          "evidence_text",
          "evidence_type",
          "source_name",
          "source_url"
        ],
        "title": "Evidence",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelOutputList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ModelOutput"
            },
            "title": "Items",
            "type": "array"
          },
          "region_id": {
            "title": "Region Id",
            "type": "string"
          }
        },
        "required": [
          "region_id",
          "items"
        ],
        "title": "ModelOutputList",
        "type": "object"
      },
...
      "RegionDetail": {
        "properties": {
          "geometry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Geometry"
              },
              {
                "type": "null"
              }
            ]
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/Metric"
            },
            "title": "Metrics",
            "type": "array"
          },
          "missing_metric_categories": {
            "description": "Metric categories with no observation for this region.",
            "items": {
              "type": "string"
            },
            "title": "Missing Metric Categories",
            "type": "array"
          },
          "region": {
            "$ref": "#/components/schemas/Region"
          }
        },
        "required": [
          "region",
          "metrics",
          "missing_metric_categories",
          "geometry"
        ],
        "title": "RegionDetail",
        "type": "object"
      },
      "RegionList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Region"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Matching regions before limit/offset.",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "RegionList",
        "type": "object"
      },
      "ScreeningResponse": {
        "properties": {
          "direction": {
            "description": "Which residuals the priority flag marks for this outcome.",
            "enum": [
              "low",
              "high",
              "absolute"
            ],
            "title": "Direction",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScreeningRow"
            },
            "title": "Items",
            "type": "array"
          },
          "outcome_name": {
            "title": "Outcome Name",
            "type": "string"
          },
          "total": {
            "description": "Regions with at least one flagged specification.",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "outcome_name",
          "direction",
          "total",
          "items"
        ],
        "title": "ScreeningResponse",
        "type": "object"
      },
...
      "SummaryCategoryList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SummaryCategory"
            },
            "title": "Items",
            "type": "array"
          },
          "region_id": {
            "title": "Region Id",
            "type": "string"
          }
        },
        "required": [
          "region_id",
          "items"
        ],
        "title": "SummaryCategoryList",
        "type": "object"
      },
      "SummaryRequest": {
        "properties": {
          "category": {
            "default": "overview",
            "examples": [
              "overview",
              "observed_der",
              "model_pv"
            ],
            "maxLength": 64,
            "title": "Category",
            "type": "string"
          },
          "region_id": {
            "examples": [
              "90001"
            ],
            "maxLength": 16,
            "title": "Region Id",
            "type": "string"
          }
        },
        "required": [
          "region_id"
        ],
        "title": "SummaryRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Region-centered access to DER adoption, demographic context, offline regression screening outputs, and precomputed evidence-grounded summaries for California ZCTAs. Every resource is keyed by a five-character region_id.",
    "title": "California DER Decision Support API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/compare": {
      "get": {
        "operationId": "compare_api_compare_get",
        "parameters": [
          {
            "description": "Repeat the parameter or pass a comma-separated list.",
            "in": "query",
            "name": "region_ids",
            "required": true,
            "schema": {
              "description": "Repeat the parameter or pass a comma-separated list.",
              "items": {
                "type": "string"
              },
              "maxItems": 12,
              "title": "Region Ids",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare",
        "tags": [
          "regions"
        ]
      }
    },
...
    "/api/model-output/{region_id}": {
      "get": {
        "description": "Stored offline regression outputs. Nothing is fit or scored per request.",
        "operationId": "get_model_output_api_model_output__region_id__get",
        "parameters": [
          {
            "description": "Five-character ZCTA.",
            "in": "path",
            "name": "region_id",
            "required": true,
            "schema": {
              "description": "Five-character ZCTA.",
              "maxLength": 16,
              "title": "Region Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "outcome_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome Name"
            }
          },
          {
            "in": "query",
            "name": "model_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelOutputList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Model Output",
        "tags": [
          "models"
        ]
      }
    },
    "/api/regions": {
      "get": {
        "operationId": "list_regions_api_regions_get",
        "parameters": [
          {
            "description": "Prefix of a ZCTA or county, or part of a name.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Prefix of a ZCTA or county, or part of a name.",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "county",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "County"
            }
          },
          {
            "description": "Utility acronym or full name.",
            "in": "query",
            "name": "utility",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Utility acronym or full name.",
              "title": "Utility"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Regions",
        "tags": [
          "regions"
        ]
      }
    },
    "/api/regions/{region_id}": {
      "get": {
        "description": "Region metadata, every metric observation, missing categories, and geometry.",
        "operationId": "get_region_api_regions__region_id__get",
        "parameters": [
          {
            "description": "Five-character ZCTA.",
            "in": "path",
            "name": "region_id",
            "required": true,
            "schema": {
              "description": "Five-character ZCTA.",
              "maxLength": 16,
              "title": "Region Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Region",
        "tags": [
          "regions"
        ]
      }
    },
    "/api/regions/{region_id}/evidence": {
      "get": {
        "operationId": "get_region_evidence_api_regions__region_id__evidence_get",
        "parameters": [
          {
            "description": "Five-character ZCTA.",
            "in": "path",
            "name": "region_id",
            "required": true,
            "schema": {
              "description": "Five-character ZCTA.",
              "maxLength": 16,
              "title": "Region Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "evidence_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Evidence Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Evidence"
                  },
                  "title": "Response Get Region Evidence Api Regions  Region Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Region Evidence",
        "tags": [
          "regions"
        ]
      }
    },
    "/api/regions/{region_id}/summaries": {
      "get": {
        "description": "Which summary categories have a stored summary for this region.",
        "operationId": "list_region_summaries_api_regions__region_id__summaries_get",
        "parameters": [
          {
            "description": "Five-character ZCTA.",
            "in": "path",
            "name": "region_id",
            "required": true,
            "schema": {
              "description": "Five-character ZCTA.",
              "maxLength": 16,
              "title": "Region Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryCategoryList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Region Summaries",
        "tags": [
          "summaries"
        ]
      }
    },
    "/api/screening": {
      "get": {
        "description": "Regions flagged as priorities, ranked by agreement across specifications.\n\nDER outcomes flag adoption below prediction (residual at or below the 25th\npercentile); burden outcomes flag burden above prediction (75th or higher).",
        "operationId": "screening_api_screening_get",
        "parameters": [
          {
            "description": "Outcome to screen, e.g. y_pv.",
            "in": "query",
            "name": "outcome_name",
            "required": false,
            "schema": {
              "default": "y_pv",
              "description": "Outcome to screen, e.g. y_pv.",
              "maxLength": 100,
              "title": "Outcome Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "county",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "County"
            }
          },
          {
            "in": "query",
            "name": "utility",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Utility"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreeningResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Screening",
        "tags": [
          "models"
        ]
      }
    },
    "/api/summaries": {
      "post": {
        "description": "Return the stored, evidence-grounded summary for a region and category.\n\nSummaries are generated offline; this endpoint never calls an LLM. When no\ntraceable summary exists, status is \"unavailable\" and the retrieved evidence\nis returned with the reason in `warnings`.",
        "operationId": "request_summary_api_summaries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Summary",
        "tags": [
          "summaries"
        ]
      }
    },
...
  }
}